+++ hpc/nvidia/cuda/samples/cublasLtMatmul.cu
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
 #include <stdio.h>
-#if defined(HAVE_HCCL)
+#if defined(HAVE_HPCC)
 # include <hc_runtime.h>
 # include <hcblas/hcblas.h>
 # include <hcblas/hcblasLt.h>