+++ hpc/nvidia/cuda/samples/cudaEvent.cu
@@ -6,7 +6,7 @@
 #include <stdbool.h>
 #include <unistd.h>
 #include <time.h>
-#if defined(HAVE_HCCL)
+#if defined(HAVE_HPCC)
 # include <hc_runtime.h>
 # include "hpcc_helpers.h"
 # include "cuda_adapter.h"