+++ hpc/nvidia/cuda/samples/cudaGetDeviceProperties.cu
@@ -19,7 +19,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <string.h>
-#if defined(HAVE_HCCL)
+#if defined(HAVE_HPCC)
 #include <hc_runtime.h>
 #include "hpcc_helpers.h"
 #include "cuda_adapter.h"