+++ CHANGELOG.md
@@ -14,10 +14,12 @@ and this project adheres to
 - Add ieee/;
 - Add init/kernel/cmdline.md;
 #### Changed
+- hpc: Rename `HAVE_HCCL` to `HAVE_HPCC`;
 - targets: don't use `LD_LIBRARY_PATH` CC PFX;
 - kconfig.mk: Move it into template/;
 #### Deprecated
 #### Removed
+- hpc: remove HAVE_HCCL;
 #### Fixed
 #### Security
 #### Docs